Extract interface from Log class
Rename Log to BuildLog, extract Log interface.
